test(worker-sessions): split HTTP handler coverage
